修改

+++ TiHMS/CommonModule/src/main/java/tiquartet/CommonModule/vo/HotelBriefVO.java
@@ -2,19 +2,19 @@ package tiquartet.CommonModule.vo;
 
 public class HotelBriefVO {
 	//酒店编号
-	public int hotelID;
+	public long hotelID;
 	//酒店星级
     public int star;
-    //客户在该酒店总订单数量
+    //
     public int numOfAllOrder;
-    //平均评分
+    //
     public int numOfEndOrder;
-    //已执行订单数量
-    public double avgragegrade;
-    //城市名
+    //酒店评分
+    public int avgragegrade;
+    //城市名称
     public String cityName;
-    //商圈名
+    //商圈名称
     public String circleName;
-    //酒店名
+    //酒店名称
     public String hotelName;
 }
